+++ keccak-air/examples/prove_goldilocks_sha256.rs
@@ -68,7 +68,7 @@ fn main() -> Result<(), impl Debug> {
     let config = MyConfig::new(pcs);
 
     let mut challenger = Challenger::from_hasher(vec![], byte_hash);
-    let proof = prove(&config, &KeccakAir {}, &mut challenger, trace, &vec![]);
+    let proof = prove(&config, &KeccakAir {}, &mut challenger, trace, &[]);
 
     let mut challenger = Challenger::from_hasher(vec![], byte_hash);
     verify(&config, &KeccakAir {}, &mut challenger, &proof, &vec![])
