+++ uni-stark/src/verifier.rs
@@ -128,7 +128,7 @@ where
     let alpha: SC::Challenge = challenger.sample_ext_element();
     challenger.observe(commitments.quotient_chunks.clone());
 
-    let zeta: SC::Challenge = challenger.sample_ext_element();
+    let zeta: SC::Challenge = challenger.sample();
     let zeta_next = trace_domain.next_point(zeta).unwrap();
 
     pcs.verify(
